🗃	데이버베이스 관련 수정 : 주문시 가격을 Decimal에서 Long으로 받는 걸로 변경
생각해보니까 한국돈으로 거래하니까 소수점이 생길 일이 없는데 소수점으로 해버림 그로인한 수정

+++ backend/src/main/resources/sql/table/table.sql
@@ -214,7 +214,7 @@ CREATE TABLE transactions (
                               transaction_id BIGSERIAL PRIMARY KEY,
                               transaction_type VARCHAR(10) NOT NULL,
                               quantity DECIMAL(20,10) NOT NULL,
-                              price DECIMAL(20,10) NOT NULL,
+                              price BIGINT NOT NULL,
                               create_date TIMESTAMP NOT NULL,
                               status varchar(20) NOT NULL,
                               execution_date TIMESTAMP NULL,
